Basic emission recalc almost done. DB is wrong, redo the script to csv with "." and not " "
The emission recall is almost done, however there are some problems
with the database structure. This should be resolved before anything
else. It seems the data is imported wrong for some cases? The numpy
array also return the wrong value on np.argmax - wrong col?? Check the
todos in the code